Utilities update + test import adjustment
