+++ tests/utilities/base.ts
@@ -1,32 +1,12 @@
-import { TestInfo, test as baseTest } from '@playwright/test';
-import { Homepage } from '../../page-objects/homepage';
-import { Login } from '../../page-objects/login';
-import { Product } from '../../page-objects/productpage';
-import { Contact } from '../../page-objects/contact';
-import { Basepage } from './actions';
+import type { Page } from '@playwright/test';
+import { BrowserContext } from '@playwright/test';
 
-const test = baseTest.extend<{
-    homepage: Homepage;
-    basepage: Basepage;
-    login: Login;
-    productpage: Product;
-    contact: Contact;
-}>({
-    homepage: async ({ page, context}, use) => {
-        await use(new Homepage(page, context));
-    },
-    basepage: async ({ page, context}, use) => {
-        await use(new Basepage(page, context));
-    },
-    login: async ({ page, context}, use) => {
-        await use(new Login(page, context));
-    },
-    productpage: async ({ page, context}, use) => {
-        await use(new Product(page, context));
-    },
-    contact: async ({ page, context}, use) => {
-        await use(new Contact(page, context));
+export class Basepage {
+    readonly page: Page;
+    readonly context: BrowserContext;
+
+    constructor(page: Page, context: BrowserContext) {
+        this.page = page;
+        this.context = context;
     }
-})
-
-export default test;
+}
